requirements.txt
requests>=2.31
lxml>=5.1
typer>=0.12
pyyaml>=6.0
python-dotenv>=1.0
tenacity>=8.2
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
